* (OS/2)「内蔵コマンド | more」のパイプを Ctrl-C で中止しても、出力が止まらない問題を修正 * (OS/2) 非PMアプリを起動する時に spawnvp ではなく、DosExecPgm を使うようにした。

+++ mysystem.h
@@ -2,7 +2,7 @@
 #define MYSYSTEM_H
 
 #ifdef OS2EMX
-    typedef int phandle_t;
+    typedef unsigned long phandle_t;
 #else
 #   include <windows.h>
     typedef HANDLE phandle_t;
